refactor: consolidate test structure by migrating integration tests to unit/
- Migrated all tests from tests/integration/ to tests/unit/ subfolders
- Created per-folder conftest.py files with proper fixtures (mocked auth,
  temp SQLite DB) for test isolation without polluting global namespace
- Removed tests/integration/ directory entirely (orphaned Vault/E2E conftest
  had no consuming test files)
- Fixed performance tests by adding stage repository mocks for
  GENERATE_INPUT_FILES and CREATE_LOCAL_REPOSITORY stages
- Updated tests/README.md to reflect consolidated structure
- Fixed stale path comments in test files
- All tests pass: 1268 passed, 22 skipped, 2 pre-existing failures
  (dependency rule violations in api/jobs/routes.py)
