+++ build_stream/tests/unit/api/build_image/test_build_image_api.py
