+++ build_stream/tests/unit/api/local_repo/test_create_local_repo_edge_cases.py
@@ -21,7 +21,7 @@ from unittest.mock import patch
 
 import pytest
 
-from tests.integration.api.local_repo.conftest import setup_input_files
+from tests.unit.api.local_repo.conftest import setup_input_files
 
 
 class TestCreateLocalRepoEdgeCases:
