+++ build_stream/tests/unit/api/parse_catalog/conftest.py
@@ -12,65 +12,58 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Shared fixtures for Validate API integration tests."""
+"""Shared fixtures for ParseCatalog API tests."""
 
 import os
-from pathlib import Path
-from typing import Dict
+from datetime import datetime, timedelta, timezone
+from typing import Dict, Generator
+from unittest.mock import patch
 
 import pytest
 
-from fastapi.testclient import TestClient
-from api.dependencies import verify_token
-
 from infra.id_generator import UUIDv4Generator
-from core.jobs.value_objects import StageState
 
 
 @pytest.fixture(scope="function")
 def client(tmp_path):
     """Create test client with fresh container for each test."""
     os.environ["ENV"] = "dev"
-    # Use file-based SQLite database for integration tests
     db_file = tmp_path / "test.db"
     db_url = f"sqlite:///{db_file}"
     os.environ["DATABASE_URL"] = db_url
-    
-    # Import app after setting DATABASE_URL
-    from main import app
+
+    from main import app  # pylint: disable=import-outside-toplevel
 
     def mock_verify_token():
         return {
             "sub": "test-client-123",
             "client_id": "test-client-123",
-            "scopes": ["job:write", "job:read"]
+            "scopes": ["job:write", "job:read", "catalog:read"]
         }
 
+    from api.dependencies import verify_token  # pylint: disable=import-outside-toplevel
     app.dependency_overrides[verify_token] = mock_verify_token
-    
-    # Create database tables before starting test client
-    from infra.db.models import Base
-    import infra.db.config as config_module
-    import importlib
-    
-    # Refresh db_config to pick up new DATABASE_URL
+
+    from infra.db.models import Base  # pylint: disable=import-outside-toplevel
+    import infra.db.config as config_module  # pylint: disable=import-outside-toplevel
+    import importlib  # pylint: disable=import-outside-toplevel
+
     config_module.db_config = config_module.DatabaseConfig()
-    
-    # Re-import session module to pick up new db_config
-    import infra.db.session
+
+    import infra.db.session  # pylint: disable=import-outside-toplevel
     importlib.reload(infra.db.session)
     session_module = infra.db.session
-    
-    from sqlalchemy import create_engine
+
+    from sqlalchemy import create_engine  # pylint: disable=import-outside-toplevel
     engine = create_engine(db_url)
-    session_module._engine = engine
-    session_module._session_factory = None
+    session_module._engine = engine  # pylint: disable=protected-access
+    session_module._session_factory = None  # pylint: disable=protected-access
     Base.metadata.create_all(engine)
-    
+
+    from fastapi.testclient import TestClient  # pylint: disable=import-outside-toplevel
     with TestClient(app) as test_client:
         yield test_client
 
-    # Cleanup
     app.dependency_overrides.clear()
 
 
@@ -106,48 +99,23 @@ def created_job(client, auth_headers) -> str:
 
 
 @pytest.fixture
-def job_with_completed_restart(client, auth_headers, created_job, monkeypatch) -> str:
-    """Create a job with a completed restart stage."""
-    from core.jobs.entities import Stage
-    from core.jobs.value_objects import JobId, StageName, StageType
-    
-    # Mock the stage repository to return a completed restart stage
-    def mock_find_by_job_and_name(self, job_id, stage_name):
-        # Handle JobId objects or string job_id
-        job_id_str = str(job_id)
-        
-        if stage_name.value == StageType.RESTART.value:
-            stage = Stage(
-                job_id=JobId(job_id_str),
-                stage_name=StageName(StageType.RESTART.value),
-                stage_state=StageState.COMPLETED,
-                attempt=1
-            )
-            return stage
-        return None
-    
-    # Apply the mock - in dev mode, it uses container's stage repository
-    from container import container
-    monkeypatch.setattr(
-        container.stage_repository().__class__,
-        "find_by_job_and_name",
-        mock_find_by_job_and_name
-    )
-    
-    return created_job
+def mock_jwt_validation() -> Generator[None, None, None]:
+    """Mock JWT validation for tests that use TestClient(app) directly.
 
+    This fixture bypasses JWT validation to allow testing of API endpoints
+    without requiring actual JWT keys.
+    """
+    with patch("api.auth.jwt_handler.JWTHandler.validate_token") as mock_validate:
+        # Mock successful token validation
+        from api.auth.jwt_handler import TokenData  # pylint: disable=import-outside-toplevel
 
-@pytest.fixture
-def nfs_queue_dir(tmp_path):
-    """Create temporary NFS queue directory structure."""
-    requests_dir = tmp_path / "requests"
-    results_dir = tmp_path / "results"
-    archive_dir = tmp_path / "archive" / "results"
-    processing_dir = tmp_path / "processing"
-
-    requests_dir.mkdir(parents=True)
-    results_dir.mkdir(parents=True)
-    archive_dir.mkdir(parents=True)
-    processing_dir.mkdir(parents=True)
-
-    return tmp_path
+        now = datetime.now(timezone.utc)
+        mock_validate.return_value = TokenData(
+            client_id="test-client",
+            client_name="test-client",
+            scopes=["catalog:read", "catalog:write"],
+            issued_at=now,
+            expires_at=now + timedelta(hours=1),
+            token_id="test-token-id",
+        )
+        yield
