+++ build_stream/tests/unit/api/validate/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unit tests for validate API module."""
+"""Unit tests for Validate API."""
