+++ build_stream/tests/unit/core/catalog/diff_input_configs.py
