+++ build_stream/tests/unit/core/catalog/test_adapter_cli_defaults.py
