+++ build_stream/tests/unit/core/catalog/test_generator_cli_defaults.py
