+++ build_stream/tests/unit/core/image_group/test_image_group_data_model.py
